Readjusting the employee's salary
a program that takes an employee's salary and readjusts it according to the following criteria, based on the current salary;
B. Salaries up to R$280.00 (including): 20% increase;
w. Salaries between R$280.00 and R$700.00: 15% increase;
d. Salaries between R$700.00 and R$1500.00: 10% increase;
It is. Salaries from R$1500.00 onwards: 5% increase
After the increase is carried out; report on screen;

The. The salary before the adjustment;
B. The percentage of increase applied;
w. The amount of the increase;
d. The new salary, after the increase.